aeee garbage.

+++ src/tcp_checksum.c
@@ -4,7 +4,7 @@
 #include "tcp.h"
 
 u16
-tcp_checksum (u16 *ptr, int nbytes) {
+tcp_checksum (u16 *ptr, u64 nbytes) {
   u32 sum = 0;
 
   while (nbytes > 1) {
